backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1

# AI — mood + age detection
google-generativeai==0.7.2

# Audio ingestion
yt-dlp==2024.5.27          # YouTube audio download (requires ffmpeg on PATH)
pydub==0.25.1              # MP3/WAV conversion (requires ffmpeg on PATH)

# Music transformation
librosa==0.10.2
soundfile==0.12.1
numpy==1.26.4

# Testing
pytest==8.2.2
pytest-flask==1.3.0
